requirements.txt
annotated-types==0.7.0
anyio==4.3.0
audioread==3.0.1
brotlipy==0.7.0
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flit_core==3.6.0
fsspec==2024.3.1
gluonnlp==0.8.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.0
graphviz==0.8.4
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==4.11.3
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
mxnet==1.7.0.post2
networkx==3.3
numba==0.59.1
numpy==1.23.1
orjson==3.10.3
packaging==22.0
pandas==2.2.2
pip==22.3.1
platformdirs==4.2.2
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
rich==13.7.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.0
sentencepiece==0.2.0
setuptools==65.6.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
textrank==0.1.2
threadpoolctl==3.5.0
tokenizers==0.13.3
torch==2.3.0
tqdm==4.64.1
transformers==4.24.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.14
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.11.0
textrank @ git+https://github.com/lovit/textrank.git@578c7fbbb22b24561b1bd4bab9a9ecc2dc9515e1